endpoints/upload-file/requirements.txt
jsonschema
pyjwt
